inject ante tx checker function
